% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_cp_model}
\alias{cfb_cp_model}
\title{College football completion probability model}
\format{
A raw vector representation of an 'xgboost' model
(\code{binary:logistic}). Features, in order: \code{down}, \code{distance},
\code{yards_to_goal}, \code{score_diff}, \code{seconds_remaining}, \code{is_home}, \code{period},
\code{passing_down}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_cp_model
}
\description{
College football completion probability model
}
\examples{
model <- xgboost::xgb.load.raw(cfb_cp_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
