% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_ep_model}
\alias{cfb_ep_model}
\title{College football expected points model}
\format{
A raw vector representation of an 'xgboost' model
(\code{multi:softprob}, 7 classes). Features, in order: \code{TimeSecsRem},
\code{yards_to_goal}, \code{distance}, \code{down_1}, \code{down_2}, \code{down_3}, \code{down_4},
\code{pos_score_diff_start}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}, built by
the \code{sportsdataverse-py} model pipeline.
}
\usage{
cfb_ep_model
}
\description{
Seven-class next-score model. Column \code{i} of the softmax output is
\code{cfb_model_manifest$ep_class_contract$class_order[i]}, worth
\code{cfb_model_manifest$ep_class_contract$point_values[i]} points; expected
points is the probability-weighted sum.
}
\examples{
model <- xgboost::xgb.load.raw(cfb_ep_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
