% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_fg_model}
\alias{cfb_fg_model}
\title{College football field goal model}
\format{
A raw vector representation of an 'xgboost' model
(\code{binary:logistic}). Features, in order: \code{yards_to_goal}, \code{era0},
\code{era1}, \code{era2}, \code{era3} (one-hot rule eras; cuts at 2006, 2013 and 2020,
see \code{cfb_model_cards$fg_model$era_contract}).
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_fg_model
}
\description{
College football field goal model
}
\examples{
model <- xgboost::xgb.load.raw(cfb_fg_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
