% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_punt_distribution}
\alias{cfb_punt_distribution}
\title{College football punt distance distribution}
\format{
A data frame; see \code{names(cfb_punt_distribution)}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_punt_distribution
}
\description{
Empirical distribution of net punt yardage by field position, used by the
fourth down decision models to score a punt.
}
\examples{
head(cfb_punt_distribution)
}
\keyword{datasets}
