% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_qbr_model}
\alias{cfb_qbr_model}
\title{College football quarterback rating model}
\format{
A raw vector representation of an 'xgboost' model
(\code{reg:squarederror}). Features, in order: \code{qbr_epa}, \code{sack_epa},
\code{pass_epa}, \code{rush_epa}, \code{pen_epa}, \code{spread}, \code{era0}, \code{era1}, \code{era2},
\code{era3}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_qbr_model
}
\description{
College football quarterback rating model
}
\examples{
model <- xgboost::xgb.load.raw(cfb_qbr_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
