% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_two_pt_model}
\alias{cfb_two_pt_model}
\title{College football two point conversion model}
\format{
A raw vector representation of an 'xgboost' model
(\code{binary:logistic}). Features, in order: \code{posteam_spread},
\code{posteam_total}, \code{pos_score_diff}, \code{era} (integer era bucket).
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_two_pt_model
}
\description{
College football two point conversion model
}
\examples{
model <- xgboost::xgb.load.raw(cfb_two_pt_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
