% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_wp_naive_model}
\alias{cfb_wp_naive_model}
\title{College football win probability model (without point spread)}
\format{
A raw vector representation of an 'xgboost' model
(\code{binary:logistic}). Same features as \link{cfb_wp_spread_model} without
\code{spread_time}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_wp_naive_model
}
\description{
College football win probability model (without point spread)
}
\examples{
model <- xgboost::xgb.load.raw(cfb_wp_naive_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
