% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_wp_spread_model}
\alias{cfb_wp_spread_model}
\title{College football win probability model (with point spread)}
\format{
A raw vector representation of an 'xgboost' model
(\code{binary:logistic}). Features, in order:
\code{pos_team_receives_2H_kickoff}, \code{spread_time}, \code{TimeSecsRem},
\code{adj_TimeSecsRem}, \code{ExpScoreDiff_Time_Ratio}, \code{pos_score_diff_start},
\code{down}, \code{distance}, \code{yards_to_goal}, \code{is_home},
\code{pos_team_timeouts_rem_before}, \code{def_pos_team_timeouts_rem_before},
\code{period}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_wp_spread_model
}
\description{
College football win probability model (with point spread)
}
\examples{
model <- xgboost::xgb.load.raw(cfb_wp_spread_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
