% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cfb_xpass_model}
\alias{cfb_xpass_model}
\title{College football expected pass model}
\format{
A raw vector representation of an 'xgboost' model
(\code{binary:logistic}). Features, in order: \code{down}, \code{distance},
\code{yards_to_goal}, \code{pos_score_diff}, \code{TimeSecsRem}, \code{era}, \code{period}.
}
\source{
The \code{cfb_model_artifacts} release of
\url{https://github.com/sportsdataverse/sportsdataverse-data}.
}
\usage{
cfb_xpass_model
}
\description{
College football expected pass model
}
\examples{
model <- xgboost::xgb.load.raw(cfb_xpass_model)
xgboost::getinfo(model, "feature_name")
}
\keyword{datasets}
